Add protection classes. Clean code.
