+++ Trust/Protection/Coordinators/SplashCoordinator.swift
@@ -1,0 +1,22 @@
+// Copyright SIX DAY LLC. All rights reserved.
+
+import UIKit
+
+class SplashCoordinator: Coordinator {
+    var coordinators: [Coordinator] = []
+    let navigationController: UINavigationController
+    lazy var rootViewController: SplashViewController = {
+        return SplashViewController()
+    }()
+    init(
+        navigationController: UINavigationController = NavigationController()
+        ) {
+        self.navigationController = navigationController
+    }
+    func start() {
+        navigationController.viewControllers = [rootViewController]
+    }
+    @objc func dismiss() {
+        navigationController.dismiss(animated: true, completion: nil)
+    }
+}
